add `listing-page` to mirage
we've been using `show` as a short hand for `listing-pages` for a while,
but they have very different use cases. For now shows are just the
things rendered on `/shows`, but `listing-pages` are a detail view on
shows, which include the `api-response` abstraction and all that other
stuff.

this commit makes that distinction more explicit. in the future, shows
will be able to fully encapsulate the sub routing, but for now, this
works.
