+++ mirage/models/listing-page.js
@@ -1,0 +1,5 @@
+import { Model, belongsTo } from 'ember-cli-mirage';
+
+export default Model.extend({
+  apiResponse: belongsTo()
+});
