+++ mirage/models/show.js
@@ -1,5 +1,4 @@
-import { Model, belongsTo } from 'ember-cli-mirage';
+import { Model } from 'ember-cli-mirage';
 
 export default Model.extend({
-  apiResponse: belongsTo()
 });
