+++ mirage/serializers/show.js
@@ -1,16 +1,5 @@
-import ApplicationSerializer from './application';
-import { dasherize } from 'ember-string';
+import { JSONAPISerializer } from 'ember-cli-mirage';
 
-export default ApplicationSerializer.extend({
-  include: ['apiResponse'],
-  keyForRelationship: key => dasherize(key),
-  typeKeyForModel(model) {
-    // This is a Mirage serialization bug, the wrong serializer
-    // is being used for 'story'.
-    if (model.modelName === 'show') {
-      return 'channel';
-    } else {
-      return ApplicationSerializer.prototype.typeKeyForModel(model);
-    }
-  },
+export default JSONAPISerializer.extend({
+  include: ['image']
 });
